tesing UPDATE of phone numbers

+++ public_html/phone_num.php
@@ -125,7 +125,7 @@
 
 				$result->fetch();
 				$result->close();
-				if ($col1 == 0 ){
+				if (if $col1 != null && $col1 == 0 ){
 					$stmt = $conn->prepare("DELETE FROM Phone WHERE phone_number = ? LIMIT 1");
 					$stmt->bind_param('s', $bad_num);
 					if($stmt->execute()) {
